requirements.txt
youtube-transcript-api
langchain
openai
pydantic
python-dotenv
requests
langgraph
langchain_community
langchain-ollama
langsmith
fpdf
fastapi
streamlit
uvicorn
openai-whisper
yt-dlp
graphviz
beautifulsoup4 
firecrawl